added top 5 command and removed not needed things
